fz.ZMCSW032D_cover_position and tz.ZMCSW032D_cover_position
